second: more work on 320x400 mode

+++ vga_emulator/second/jp/sin1024.h
@@ -1,1 +1,1 @@
-extern signed int sin1024[1024];
+extern int16_t sin1024[1024];
